+++ publish.sh
@@ -116,7 +116,7 @@ local HASH=$(
 echo $HASH
 }
 
-
+# Function to check if any source files in the directory have been changed
 haschanged() {
   local dir=$1
   local checksum_file="${dir}/.checksum"
@@ -136,7 +136,6 @@ haschanged() {
       return 1  # False, the directory has not changed
   fi
 }
-
 update_checksum() {
   local dir=$1
   local checksum_file="${dir}/.checksum"
@@ -148,6 +147,36 @@ update_checksum() {
   echo "$current_checksum" > "$checksum_file"
 }
 
+# Function to check if the submodule commit hash has changed
+hassubmodulechanged() {
+    local dir=$1
+    local hash_file="${dir}/.commit-hash"
+    # Get the current commit hash of the submodule
+    cd "$dir" || exit 1
+    current_hash=$(git rev-parse HEAD)
+    cd - > /dev/null || exit 1
+    # Check if the hash file exists and read the previous hash
+    if [ -f "$hash_file" ]; then
+        previous_hash=$(cat "$hash_file")
+    else
+        previous_hash=""
+    fi
+    if [ "$current_hash" != "$previous_hash" ]; then
+        return 0  # True, the submodule has changed
+    else
+        return 1  # False, the submodule has not changed
+    fi
+}
+update_submodule_hash() {
+    local dir=$1
+    local hash_file="${dir}/.commit-hash"
+    # Get the current commit hash of the submodule
+    cd "$dir" || exit 1
+    current_hash=$(git rev-parse HEAD)
+    cd - > /dev/null || exit 1
+    # Save the current hash
+    echo "$current_hash" > "$hash_file"
+}
 
 dir=lma-browser-extension-stack
 cd $dir
@@ -251,6 +280,18 @@ else
 echo "SKIPPING $dir (unchanged)"
 fi
 
+dir=lma-bedrockagent-stack
+if haschanged $dir; then
+echo "PACKAGING $dir"
+pushd $dir
+chmod +x ./publish.sh
+./publish.sh $BUCKET $PREFIX_AND_VERSION $REGION || exit 1
+popd
+update_checksum $dir
+else
+echo "SKIPPING $dir (unchanged)"
+fi
+
 dir=lma-websocket-transcriber-stack
 if haschanged $dir; then
 echo "PACKAGING $dir"
@@ -281,7 +322,6 @@ dir=lma-llm-template-setup-stack
 if haschanged $dir; then
 echo "PACKAGING $dir/deployment"
 pushd $dir/deployment
-
 # by hashing the contents of the source folder, we can force the custom resource lambda to re-run
 # when the code or prompt template contents change.
 echo "Computing hash of src folder contents"
@@ -308,43 +348,26 @@ else
 echo "SKIPPING $dir (unchanged)"
 fi
 
-echo "Initialize and update git submodules"
-git submodule init
-git submodule update
-
-dir=submodule-aws-qnabot-plugins
-if haschanged $dir; then
-echo "PACKAGING $dir"
-pushd $dir
-chmod +x ./publish.sh
-./publish.sh $BUCKET $PREFIX_AND_VERSION/aws-qnabot-plugins || exit 1
-popd
-update_checksum $dir
-else
-echo "SKIPPING $dir (unchanged)"
-fi
-
 dir=submodule-aws-qnabot
-if haschanged $dir; then
-echo "PACKAGING $dir"
+echo "UPDATING $dir"
 git submodule init
+echo "Removing any QnAbot changes from previous builds"
+pushd $dir && git checkout . && popd
 git submodule update
-echo "Applying patch files to simplify UX by removing some QnABot options not needed for lma"
 # lma customizations
-cp -v ./patches/qnabot/Makefile $dir/Makefile
-cp -v ./patches/qnabot/templates_examples_examples_index.js $dir/templates/examples/examples/index.js
-cp -v ./patches/qnabot/templates_examples_extensions_index.js $dir/templates/examples/extensions/index.js
+echo "Applying patch files to remove unused KMS keys from QnABot and customize designer settings page"
+cp -v ./patches/qnabot/templates_examples_examples_index.js $dir/source/templates/examples/examples/index.js
+cp -v ./patches/qnabot/templates_examples_extensions_index.js $dir/source/templates/examples/extensions/index.js
+cp -v ./patches/qnabot/website_js_lib_store_api_actions_settings.js $dir/source/website/js/lib/store/api/actions/settings.js
 echo "modify QnABot version string from 'N.N.N' to 'N.N.N-lma'"
 # Detection of differences. sed varies betwen GNU sed and BSD sed
 if sed --version 2>/dev/null | grep -q GNU; then # GNU sed
-  sed -i 's/"version": *"\([0-9]*\.[0-9]*\.[0-9]*\)"/"version": "\1-lma"/' $dir/package.json
+  sed -i 's/"version": *"\([0-9]*\.[0-9]*\.[0-9]*\)"/"version": "\1-lma"/' $dir/source/package.json
 else # BSD like sed
-  sed -i '' 's/"version": *"\([0-9]*\.[0-9]*\.[0-9]*\)"/"version": "\1-lma"/' $dir/package.json
+  sed -i '' 's/"version": *"\([0-9]*\.[0-9]*\.[0-9]*\)"/"version": "\1-lma"/' $dir/source/package.json
 fi
-pushd $dir
-rm -fr ./ml_model/llm-qa-summarize # remove deleted folder if left over from previous build.
-mkdir -p build/templates/dev
-cat > config.json <<_EOF
+echo "Creating config.json"
+cat > $dir/source/config.json <<_EOF
 {
   "profile": "${AWS_PROFILE:-default}",
   "region": "${REGION}",
@@ -353,11 +376,24 @@ cat > config.json <<_EOF
   "noStackOutput": true
 }
 _EOF
+
+# only re-build QnABot if patch files or submodule version has changed
+if haschanged ./patches/qnabot || hassubmodulechanged $dir; then
+
+echo "PACKAGING $dir"
+
+pushd $dir/source
+mkdir -p build/templates/dev
 npm install
 npm run build || exit 1
+# Rename OpensearchDomain resource in template to force resource replacement during upgrade/downgrade
+# If the resource name is not changed, then CloudFomration does an inline upgrade from OpenSearch 1.3 to 2.1, but this upgrade cannot be reversed
+# which can create a problem with ROLLBACK if there is a stack failure during the upgrade.
+cat ./build/templates/master.json | sed -e "s%OpensearchDomain%LMAQnaBotOpensearchDomain%g" > ./build/templates/qnabot-main.json
 aws s3 sync ./build/ s3://${BUCKET}/${PREFIX_AND_VERSION}/aws-qnabot/ --delete 
 popd
-update_checksum $dir
+update_checksum ./patches/qnabot
+update_submodule_hash $dir
 else
 echo "SKIPPING $dir (unchanged)"
 fi
@@ -371,6 +407,7 @@ echo "   <ARTIFACT_PREFIX_TOKEN> with prefix: $PREFIX_AND_VERSION"
 echo "   <VERSION_TOKEN> with version: $VERSION"
 echo "   <REGION_TOKEN> with region: $REGION"
 echo "   <BROWSER_EXTENSION_SRC_S3_LOCATION_TOKEN> with public: $BROWSER_EXTENSION_SRC_S3_LOCATION"
+echo "   <VIRTUAL_PARTICIPANT_SRC_S3_LOCATION_TOKEN> with public: $VIRTUAL_PARTICIPANT_SRC_S3_LOCATION"
 cat ./$MAIN_TEMPLATE | 
 sed -e "s%<ARTIFACT_BUCKET_TOKEN>%$BUCKET%g" | 
 sed -e "s%<ARTIFACT_PREFIX_TOKEN>%$PREFIX_AND_VERSION%g" |